`EnforceBytecodeVersion` fails with "Restricted to JDK 11 yet […] contains […] targeted to JDK 17"
